Move $spam to chatcommands.js
