+++ lib/chatcommands.js
@@ -10,7 +10,7 @@ import {Rank, RESTART_TIMES_FILE_NAME} from './constants.js';
 import {handle as customHandle} from './custom.js';
 import {auditLog, errorLog, infoLog} from './logger.js';
 import {sendHybridModPermissions} from './permissions.js';
-import {filterMsg, genericUIDLoop, getCurrentUnixTimestamp, getUser, kill, parseBumpData, parseDeleteData, parseForecastData, parseMediaLink, parseUserlimit, sendPmsWithRateLimit, sleep} from './utils.js';
+import {filterMsg, genericUIDLoop, getCurrentUnixTimestamp, getUser, kill, parseBumpData, parseDeleteData, parseForecastData, parseMediaLink, parseUserlimit, sendMessagesWithRateLimit, sendPmsWithRateLimit, sleep} from './utils.js';
 import {validateYouTubeVideo} from './validate.js';
 
 const execAsync = promisify(exec);
@@ -1385,6 +1385,27 @@ CHAT_HANDLERS.set('skip', async (bot, username, msg) => {
   }
 
   bot.deleteVideo(bot.currentUID);
+});
+
+CHAT_HANDLERS.set('spam', async (bot, username, msg) => {
+  if (!msg) {
+    return;
+  }
+  if (!(await bot.checkPermission(username, Rank.MOD, null))) {
+    bot.sendChatMsg(`${username} does not have permission to spam. FeelsWeirdMan`);
+    return;
+  }
+
+  const nMsg = msg.split(' ')[0];
+  const n = parseInt(nMsg, 10);
+  if (isNaN(n)) {
+    bot.sendChatMsg(`Failed to parse spam times. Example: $spam 5 PeepoGlad`);
+    return;
+  }
+
+  const rest = msg.slice(2);
+
+  await sendMessagesWithRateLimit(bot, Array(n).fill(rest));
 });
 
 // Shows basic database stats.
