+++ lib/custom.js
@@ -5,9 +5,6 @@
 
 import random from 'random';
 
-import {Rank} from './constants.js';
-import {sendMessagesWithRateLimit} from './utils.js';
-
 /** @typedef {import('./cytubebot.js').CytubeBot} CytubeBot */
 
 /**
@@ -73,27 +70,6 @@ CUSTOM_HANDLERS.set('pyramid', async (bot, username, msg) => {
   bot.sendChatMsg(`⠀ ${word} ${word} ${word}`);
   bot.sendChatMsg(`⠀ ${word} ${word}`);
   bot.sendChatMsg(`⠀ ${word}`);
-});
-
-CUSTOM_HANDLERS.set('spam', async (bot, username, msg) => {
-  if (!msg) {
-    return;
-  }
-  if (!(await bot.checkPermission(username, Rank.MOD, null))) {
-    bot.sendChatMsg(`${username} does not have permission to spam. FeelsWeirdMan`);
-    return;
-  }
-
-  const nMsg = msg.split(' ')[0];
-  const n = parseInt(nMsg, 10);
-  if (isNaN(n)) {
-    bot.sendChatMsg(`Failed to parse spam times. Example: $spam 5 PeepoGlad`);
-    return;
-  }
-
-  const rest = msg.slice(2);
-
-  await sendMessagesWithRateLimit(bot, Array(n).fill(rest));
 });
 
 CUSTOM_HANDLERS.set('tuck', (bot, username, msg) => {
